+++ lib/modules/console/browser_console.js
@@ -1,29 +1,7 @@
 "use strict";
 
 const utils = require('../../utils');
-
-const LogType = {
-    log: 'log',
-    debug: 'debug',
-    info: 'info',
-    error: 'error',
-    warning: 'warning',
-    trace: 'trace'
-};
-
-class Log {
-    constructor(consoleMessage) {
-        this.message = consoleMessage;
-    }
-
-    get text() {
-        return this.message.text();
-    }
-
-    get type() {
-        return this.message.type();
-    }
-}
+const Log = require('../../models/log');
 
 module.exports = class BrowserConsole {
     constructor(browser) {
@@ -38,7 +16,7 @@ module.exports = class BrowserConsole {
     }
 
     get LogType() {
-        return LogType;
+        return Log.LogType;
     }
 
     all() {
