+++ lib/modules/cookies/browser_cookies.js
