+++ lib/modules/requests/browser_requests.js
@@ -1,7 +1,7 @@
 "use strict";
 
-const RequestFilter = require('../request_filter');
-const RequestMocker = require('../request_mocker');
+const RequestFilter = require('../../request_filter');
+const RequestMocker = require('../../request_mocker');
 
 module.exports = class BrowserRequests {
     constructor(browser) {
