+++ lib/modules/webworkers/browser_webworkers.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const WebWorker = require('../models/webworker');
+const WebWorker = require('../../models/webworker');
 
 module.exports = class BrowserWebWorkers {
     constructor(browser) {
