+++ lib/wendigo.js
@@ -18,20 +18,20 @@ const defaultSettings = {
 
 const defaultPlugins = [{
     name: "cookies",
-    plugin: require('./modules/browser_cookies')
+    plugin: require('./modules/cookies/browser_cookies')
 }, {
     name: "localStorage",
     plugin: require('./modules/local_storage/browser_local_storage'),
     assertions: require('./modules/local_storage/local_storage_assertions')
 }, {
     name: "requests",
-    plugin: require('./modules/browser_requests')
+    plugin: require('./modules/requests/browser_requests')
 }, {
     name: "console",
     plugin: require('./modules/console/browser_console')
 }, {
     name: "webworkers",
-    plugin: require('./modules/browser_webworkers')
+    plugin: require('./modules/webworkers/browser_webworkers')
 }];
